backend/requirements.txt
# App
django-cors-headers==2.0.0
django-dotenv==1.4.1
django-redis==4.7.0
django-rq==0.9.5
django==1.11.0
djangorestframework==3.6.2
psycopg2==2.7.1

# Tests
pytest
pytest-django
flake8



